requirements.txt
Django==1.11.6
djangorestframework==3.7.1
drfapikey==0.0.3
olefile==0.44
Pillow==4.3.0
psycopg2==2.7.3.2
pytz==2017.3
